lanchain-openai-examples/service-principal/requirements.txt
langchain-openai
python-dotenv
msal